Sum real numbers and delete message "Hello World!"

+++ main.cpp
@@ -4,12 +4,12 @@ using namespace std;
 
 int main()
 {
-  int a, b;
+  float a, b;
   cout << "Input first number: ";
   cin >> a;
   cout << "Input second number: ";
   cin >> b;
-  int c = a + b;
+  float c = a + b;
   cout << "Sum of numbers = " << c << endl;
 
   cin.get();
